[CALCITE-3478] Restructure tests for materialized views
